feat: Add ValueX7, scale up all supply and demand values to fix rounding error (#108)
* reproduce

* Fix rounding error test

* Add manual test for reproducing

* Fix test_manual

* I think fix, scale all amounts by MPS

* passes manual and invariant

* Add ValueX7 and MPSLib

* Switch back to uint256

* pre-commit

* Remove old references to AuctionStepLib.MPS

* Clean up

* Scale down after fullMulDiv in _accountPartiallyFilledCheckpoints

* pre-commit

* Rename mul and div to mulUint256 and divUint256

* pre-commit

* Fix order of scaleDown and remove unused parenthesis

* Remove redundant initializations in constructor

* Change to scaleByMps

* Use totalSupplyX7 internally

* pre-commit

* feedback

* fmt and pre-commit

* Add startCumulativeMps to Bid to simplify

* Simplfiy _calculateFill and add mpsRemainingInAuction to Bid library

* Use the uint256 totalSupply value instead of the scaled up X7 value

* Implement fullMulDiv for ValueX7

* chore: doc gen

* chore: doc gen

* Change constructor param to uint128

* precommit

* Add to prettierignore

* doc gen

* remove exclude from precommit

* re-add exlcude

* revert back to uint256

* Rm whitespace

---------

+++ src/AuctionFactory.sol
@@ -12,7 +12,7 @@ contract AuctionFactory is IAuctionFactory {
     address public constant USE_MSG_SENDER = 0x0000000000000000000000000000000000000001;
     /// @inheritdoc IDistributionStrategy
 
-    function initializeDistribution(address token, uint128 amount, bytes calldata configData, bytes32 salt)
+    function initializeDistribution(address token, uint256 amount, bytes calldata configData, bytes32 salt)
         external
         returns (IDistributionContract distributionContract)
     {
