+++ test/utils/MockValidationHook.sol
@@ -7,7 +7,7 @@ contract MockValidationHook is IValidationHook {
     function validate(
         uint256 maxPrice,
         bool exactIn,
-        uint128 amount,
+        uint256 amount,
         address owner,
         address sender,
         bytes calldata hookData
